tests: add lora tester mixin for modular and migrate minimax h3.
