+++ tests/modular_pipelines/testing_utils/__init__.py
@@ -5,6 +5,7 @@ from .common import (
 )
 from .guider import ModularGuiderTesterMixin
 from .loading import ModularLoadingTesterMixin
+from .lora import ModularLoraMemoryTesterMixin, ModularLoraTesterMixin
 from .memory import (
     ModularAutoOffloadTesterMixin,
     ModularGroupOffloadTesterMixin,
@@ -22,6 +23,8 @@ __all__ = [
     "ModularGroupOffloadTesterMixin",
     "ModularGuiderTesterMixin",
     "ModularLoadingTesterMixin",
+    "ModularLoraMemoryTesterMixin",
+    "ModularLoraTesterMixin",
     "ModularMemoryTesterMixin",
     "ModularOffloadTesterMixin",
     "ModularPipelineTesterMixin",
